Check and set /etc/hosts on startup
